Fix view tests
