+++ .isort.cfg
@@ -1,0 +1,12 @@
+[settings]
+combine_as_imports = true
+default_section = THIRDPARTY
+include_trailing_comma = false
+line_length = 79
+multi_line_output = 5
+skip = env,node_modules
+skip_glob = **/migrations/**
+not_skip = __init__.py
+known_django=django
+known_first_party=brouwers
+sections=FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
